Fix non-existant URL.
It used to give an invalid HTTP response, but I think what's happening
is that the server is configured to send files ending .html as
pre-gzipped, but NEGLECTS TO SEND THE Encoding: header. So user agents
get confused and think it's plain-text, when it's clearly not.
So, drop the .html extension and the server just thinks it's garbage and
deals with it anyway.